Use js module to share function across pages
