+++ public/script/authenticate.js
@@ -1,37 +1,4 @@
-function clearPopupMessage(){
-    const loginResultMessage = document.getElementById("login_result_message");
-    const signupResultMessage = document.getElementById("signup_result_message");
-    loginResultMessage.style.display = "none";
-    signupResultMessage.style.display = "none";
-}
-function clearInput(){
-    document.getElementById("login_email").value = "";
-    document.getElementById("login_password").value = "";
-    document.getElementById("signup_name").value = "";
-    document.getElementById("signup_email").value = "";
-    document.getElementById("signup_password").value = "";
-}
-function switchPopupDisplay(show){
-    const popup = document.getElementById("popup")
-    popup.style.display = show?"none":"block"
-    clearPopupMessage();
-    clearInput();
-    setPopupBoxesToDefaultDisplay();
-}
-function setPopupBoxesToDefaultDisplay(){
-    const loginBox = document.getElementById("login_box");
-    const signupBox = document.getElementById("signup_box");
-    signupBox.style.display = "none";
-    loginBox.style.display = "block";
-}
-function switchPopupBox(haveAccount){
-    const loginBox = document.getElementById("login_box");
-    const signupBox = document.getElementById("signup_box");
-    loginBox.style.display = haveAccount?"block":"none";
-    signupBox.style.display = haveAccount?"none":"block";
-    clearPopupMessage();
-    clearInput();
-}
+import { switchPopupDisplay, authenticateLogin, switchPopupBox } from "./auth.js"
 function setAccountBtn(authenticated){
     const loginBtn = document.getElementById("login_btn")
     const logoutBtn = document.getElementById("logout_btn")
@@ -119,23 +86,7 @@ function requestLogin(){
         })
     })
 }
-function requestAuthenticate(token){
-    return new Promise((resolve,reject)=>{
-        url = "/api/user/auth"
-        let head = {
-            "Authorization":`Bearer ${token}`,
-        }
-        fetch(url,{ headers : head })
-        .then((data)=>data.json())
-        .then((data)=>{
-            return resolve(data);
-        })
-        .catch(e=>{
-            console.error(e);
-            return reject(e);
-        })
-    })
-}
+
 async function signup(){
     if (checkInputNotEmpty(false)){
         let data = await requestCreateMember(); 
@@ -161,20 +112,30 @@ async function login(){
         }
     }
 }
-async function authenticateLogin(){
-    let token = localStorage.getItem("token")
-    if (token){
-        let data = await requestAuthenticate(token);
-        let result = data["data"];
-        let authenticated = result!==null?true:false;
-        setAccountBtn(authenticated);
+
+async function setLoginBtn(){
+    let authenticated = await authenticateLogin();
+    setAccountBtn(authenticated);
+}
+
+async function checkLoginForBooking(){
+    let authenticated = await authenticateLogin();
+    if (authenticated){
+        window.location.href = "/booking"
     }else{
-        setAccountBtn(false)
+        switchPopupDisplay(false);
     }
 }
+
 function logout(){
     localStorage.removeItem("token");
     location.reload();
 }
 
-authenticateLogin()
+window.signup = signup
+window.logout = logout
+window.login = login
+window.checkLoginForBooking = checkLoginForBooking
+window.switchPopupDisplay = switchPopupDisplay
+window.switchPopupBox = switchPopupBox
+setLoginBtn()
