+++ templates/index.html
@@ -8,8 +8,8 @@
     <script src="/components/loginPopup.js" type="text/javascript" defer></script>
     <script src="/components/header.js" type="text/javascript" defer></script>
     <script src="/components/footer.js" type="text/javascript" defer></script>
-    <script src="/script/main.js" defer></script>
-    <script src="/script/authenticate.js" defer></script>
+    <script type="module" src="/script/main.js" defer></script>
+    <script type="module" src="/script/authenticate.js" defer></script>
     <title>台北一日遊</title>
 </head>
 <body>
